Trust the browser to cache idol layers, don't store renders locally.
Reduces memory usage and might fix a crash that @EuricaeriS has been
having. We may make idols try to cache renders in local storage at some
point, but for now this introduces very little performance penalty On My
Machine so I'm happy with it while we wait to see if the crash keeps
happening.