Fix some clustering in our seeded RNG.
This fucks all of your saves, sorry, but it was causing some really
serious clustering around 0.2 when seeded with a small number. As a
result, most single-idol units had names beginning with E or F.

By doing one generation before using any of the numbers, we make stuff
spread out a lot more.