Remove a redundant celebration from after a battle in the story.
We already celebrate at the end of story battles, although I understand
this wasn't the case when this chapter was being written. I also am no
longer sure that's the intention; if @euricaeris has thoughts on whether
all story battles should be celebrated if won, I'd love to hear them.