Don't rerender thumbnails every time we rerender the idol list.
We used to cache full-size ones, too, but it caused OOM crashes on some
phones. Instead, though, rerendering all the thumbnails was making
toggling favourite or unit membership status pretty frustrating. Now
that they're cached, it's a lot quicker, and we're still not using
nearly as much memory as we would be if we cached all images.

We'll see if it starts crashing on old, bad phones now (im looking at
you, @euricaeris).