Let Github build everything.
I am certain that this is going to cause A Problem. If nothing else,
it's gonna take like an hour to run this one job, which probably won't
be allowed. Even if it is, though, it kinda defeats the purpose of this
script; this should probably be adapted to not resize or optimise when
running in CI.